Updated[raycast]: - Seperate [initializer] & [raycast](which has find_hit_point functions) & [hook]
Updated[cub3D.h]:
- Add more data to main header

Tested[raycast: hit_point]:
- with the updated cub3D.h and other new header files
- with mock_map.h, which should be removed after having correct parsing
