+++ include/mock_map.h
@@ -1,0 +1,9 @@
+#ifndef MOCK_MAP_H
+#define MOCK_MAP_H
+
+# define mapWidth 24
+# define mapHeight 26
+
+extern int worldMap[mapHeight][mapWidth];
+
+#endif
